Adds example marimo application and dockerfile
Adds a marimo application example to demonstrate querying data from QuixLake and visualizing it.

Includes a Dockerfile to facilitate easy deployment of the application.
The Dockerfile sets up a Python environment with the necessary dependencies,
copies the application files, and configures marimo to run in production mode with a dark theme.